[KTAI-7] ref: applied review comments
