+++ package.json
@@ -16,7 +16,6 @@
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
     "framer-motion": "^11.16.2",
-    "keytrading-ui": "file:",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   },
@@ -41,13 +40,13 @@
     "eslint-plugin-react-refresh": "^0.4.16",
     "eslint-plugin-simple-import-sort": "^12.1.0",
     "globals": "^15.14.0",
-    "husky": "^9.1.7",
+    "husky": "^9.0.11",
     "prettier": "^3.3.2",
     "stylelint": "^16.6.1",
     "stylelint-config-standard": "^36.0.1",
     "stylelint-order": "^6.0.4",
     "stylelint-semantic-groups": "^1.2.1",
-    "typescript": "^5.1.0-dev.20230515",
+    "typescript": "~5.6.2",
     "typescript-eslint": "^8.18.2",
     "vite": "^6.0.5"
   }
