+++ src/modules/Header/features/AuthedUserCard/index.tsx
@@ -1,0 +1,17 @@
+import { UserInfo } from '@modules/Header/features/UserInfo';
+
+import styles from './style.module.css';
+
+interface AuthedUserCardProps {
+  username: string;
+  avatar: string;
+}
+
+export const AuthedUserCard = ({ username, avatar }: AuthedUserCardProps) => {
+  return (
+    <div className={styles.user}>
+      <UserInfo username={username} avatar={avatar} />
+      <p className={styles.username}>{username}</p>
+    </div>
+  );
+};
