+++ src/modules/Header/features/AuthedUserCard/style.module.css
@@ -1,0 +1,10 @@
+.user {
+  display: flex;
+  gap: 10px;
+  align-items: center;
+}
+
+.username {
+  color: #333;
+  font-size: 16px;
+}
