--- src/modules/Header/features/GetStartedButton/style.module.css
@@ -1,9 +1,0 @@
-.button {
-  padding: 12px 24px;
-
-  color: #9b0000 !important;
-}
-
-.button:hover {
-  background-color: rgb(225 255 0) !important;
-}
