+++ src/modules/Header/style.module.css
@@ -16,18 +16,7 @@
   font-size: 24px;
 }
 
-.rightsection {
+.info {
   display: flex;
   align-items: center;
-}
-
-.userinfo {
-  display: flex;
-  gap: 10px;
-  align-items: center;
-}
-
-.username {
-  color: #333;
-  font-size: 16px;
 }
